cast data to float

+++ FebLaunch/DFRobot_BMX160.cpp
@@ -252,25 +252,25 @@ void DFRobot_BMX160::getAllData(struct bmx160SensorData *magn, struct bmx160Sens
     // put your main code here, to run repeatedly:
     readReg(BMX160_MAG_DATA_ADDR, data, 23);
     if(magn){
-        magn->x = (int16_t) ((data[1] << 8) | data[0]);
-        magn->y = (int16_t) ((data[3] << 8) | data[2]);
-        magn->z = (int16_t) ((data[5] << 8) | data[4]);
+        magn->x = (float) ((data[1] << 8) | data[0]);
+        magn->y = (float) ((data[3] << 8) | data[2]);
+        magn->z = (float) ((data[5] << 8) | data[4]);
         magn->x *= BMX160_MAGN_UT_LSB;
         magn->y *= BMX160_MAGN_UT_LSB;
         magn->z *= BMX160_MAGN_UT_LSB;
     }
     if(gyro){
-        gyro->x = (int16_t) ((data[9] << 8) | data[8]);
-        gyro->y = (int16_t) ((data[11] << 8) | data[10]);
-        gyro->z = (int16_t) ((data[13] << 8) | data[12]);
+        gyro->x = (float) ((data[9] << 8) | data[8]);
+        gyro->y = (float) ((data[11] << 8) | data[10]);
+        gyro->z = (float) ((data[13] << 8) | data[12]);
         gyro->x *= gyroRange;
         gyro->y *= gyroRange;
         gyro->z *= gyroRange;
     }
     if(accel){
-        accel->x = (int16_t) ((data[15] << 8) | data[14]);
-        accel->y = (int16_t) ((data[17] << 8) | data[16]);
-        accel->z = (int16_t) ((data[19] << 8) | data[18]);
+        accel->x = (float) ((data[15] << 8) | data[14]);
+        accel->y = (float) ((data[17] << 8) | data[16]);
+        accel->z = (float) ((data[19] << 8) | data[18]);
         accel->x *= accelRange;
         accel->y *= accelRange;
         accel->z *= accelRange;
